Resuelve filtrado factura x fecha

+++ src/model/IVA.java
@@ -10,7 +10,7 @@ public class IVA extends Impuestos {
     }
 
 
-    public Float calcularImpuestoIVA(String cuitProveedor, Float monto, String tipoImpuesto) {
+    public double calcularImpuestoIVA(String cuitProveedor, double monto, String tipoImpuesto) {
         Proveedor proveedor = ProveedorController.getInstancia().buscarProveedor(cuitProveedor);
 
         // Calcular impuesto IIBB solo si el proveedor existe y tiene un certificado de no retención válido para IIBB
@@ -18,16 +18,12 @@ public class IVA extends Impuestos {
             for (CertificadoDeNoRetencion certificado : proveedor.getCertificados()) {
                 if (certificado.getImpuesto() != null && tipoImpuesto.equals(certificado.getImpuesto().getTipoImpuesto())) {
                     System.out.println("Existe proveedor y certificado válido");
-                    return 0f;
+                    return 0.0;
                 }
             }
         }
-
         // Si no se cumple ninguna condición, asigna el monto directo
         System.out.println("Proveedor no encontrado o certificado de no retención no válido para IIBB");
-        return monto * 0.21f;
+        return monto * 0.21;
     }
-
-
-
 }
